+++ NEWS.md
@@ -1,11 +1,20 @@
 # pivotea release news
 
+# pivotea 1.0.2
+
+* 2024-07-15
+
+* Remove duplicated data in `hogwarts`. 
+* Add argument rm_empty_df in `pivot()`.
+* Use pkgdown.
+
 # pivotea 1.0.1
 
 * 2023-07-15
 
 * Add data: `hogwarts` which includes timetable in Hogwarts School of Witchcraft and Wizardry. 
 * Fix bug in `split_force()`.
+
 
 # pivotea 1.0.0